requirements.txt
pystray>=0.19.4
Pillow>=10.0.0
psutil>=5.9.5
